+++ internal/sdkprovider/service/cassandra/cassandra.go
@@ -1,7 +1,6 @@
 package cassandra
 
 import (
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/customdiff"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
 	"github.com/aiven/terraform-provider-aiven/internal/schemautil"
@@ -43,30 +42,7 @@ func ResourceCassandra() *schema.Resource {
 		ReadContext:   schemautil.ResourceServiceRead,
 		UpdateContext: schemautil.ResourceServiceUpdate,
 		DeleteContext: schemautil.ResourceServiceDelete,
-		CustomizeDiff: customdiff.Sequence(
-			schemautil.SetServiceTypeIfEmpty(schemautil.ServiceTypeCassandra),
-			schemautil.CustomizeDiffDisallowMultipleManyToOneKeys,
-			customdiff.IfValueChange("tag",
-				schemautil.TagsShouldNotBeEmpty,
-				schemautil.CustomizeDiffCheckUniqueTag,
-			),
-			customdiff.IfValueChange("disk_space",
-				schemautil.DiskSpaceShouldNotBeEmpty,
-				schemautil.CustomizeDiffCheckDiskSpace,
-			),
-			customdiff.IfValueChange("additional_disk_space",
-				schemautil.DiskSpaceShouldNotBeEmpty,
-				schemautil.CustomizeDiffCheckDiskSpace,
-			),
-			customdiff.IfValueChange("service_integrations",
-				schemautil.ServiceIntegrationShouldNotBeEmpty,
-				schemautil.CustomizeDiffServiceIntegrationAfterCreation,
-			),
-			customdiff.Sequence(
-				schemautil.CustomizeDiffCheckPlanAndStaticIpsCannotBeModifiedTogether,
-				schemautil.CustomizeDiffCheckStaticIPDisassociation,
-			),
-		),
+		CustomizeDiff: schemautil.CustomizeDiffGenericService(schemautil.ServiceTypeCassandra),
 		Importer: &schema.ResourceImporter{
 			StateContext: schema.ImportStatePassthroughContext,
 		},